rename controller files according to the class name
